+++ docs/tool/config.js
@@ -6,15 +6,6 @@ const path = require( 'path' );
 
 const root = path.resolve( __dirname, '../../' );
 
-// These are packages published to NPM as their own node modules.
-const npmReadyPackages = glob( 'packages/*/package.json' )
-	.map( ( fileName ) => fileName.split( '/' )[ 1 ] );
-
-// These are internal-only packages (for now), not yet published as standalone node modules.
-const gutenbergPackages = [
-	'edit-post',
-];
-
 module.exports = {
 	componentPaths: glob( 'packages/components/src/*/**/README.md' ),
 	dataNamespaces: {
@@ -36,8 +27,8 @@ module.exports = {
 		},
 		'core/edit-post': {
 			title: 'The Editor’s UI Data',
-			selectors: [ path.resolve( root, 'edit-post/store/selectors.js' ) ],
-			actions: [ path.resolve( root, 'edit-post/store/actions.js' ) ],
+			selectors: [ path.resolve( root, 'packages/edit-post/src/store/selectors.js' ) ],
+			actions: [ path.resolve( root, 'packages/edit-post/src/store/actions.js' ) ],
 		},
 		'core/nux': {
 			title: 'The NUX (New User Experience) Data',
@@ -52,16 +43,8 @@ module.exports = {
 	},
 	dataDocsOutput: path.resolve( __dirname, '../data' ),
 
-	packages: {
-		...npmReadyPackages.reduce( ( memo, packageName ) => {
-			memo[ packageName ] = { isNpmReady: true };
-			return memo;
-		}, {} ),
-		...gutenbergPackages.reduce( ( memo, packageName ) => {
-			memo[ packageName ] = { isNpmReady: false };
-			return memo;
-		}, {} ),
-	},
+	packageFileNames: glob( 'packages/*/package.json' )
+		.map( ( fileName ) => fileName.split( '/' )[ 1 ] ),
 
 	rootManifest: path.resolve( __dirname, '../root-manifest.json' ),
 	manifestOutput: path.resolve( __dirname, '../manifest.json' ),