+++ packages/block-library/src/button/index.js
@@ -32,7 +32,7 @@ const blockAttributes = {
 		attribute: 'title',
 	},
 	text: {
-		source: 'rich-text',
+		source: 'html',
 		selector: 'a',
 	},
 	backgroundColor: {