+++ packages/block-library/src/paragraph/index.js
@@ -17,7 +17,10 @@ import {
 	RichText,
 } from '@wordpress/editor';
 import { getPhrasingContentSchema } from '@wordpress/blocks';
-import { create, concat } from '@wordpress/rich-text';
+import {
+	Path,
+	SVG,
+} from '@wordpress/components';
 
 /**
  * Internal dependencies
@@ -30,8 +33,9 @@ const supports = {
 
 const schema = {
 	content: {
-		source: 'rich-text',
+		source: 'html',
 		selector: 'p',
+		default: '',
 	},
 	align: {
 		type: 'string',
@@ -70,7 +74,7 @@ export const settings = {
 
 	description: __( 'Add some basic text.' ),
 
-	icon: <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24"><path d="M11 5v7H9.5C7.6 12 6 10.4 6 8.5S7.6 5 9.5 5H11m8-2H9.5C6.5 3 4 5.5 4 8.5S6.5 14 9.5 14H11v7h2V5h2v16h2V5h2V3z" /></svg>,
+	icon: <SVG xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24"><Path d="M11 5v7H9.5C7.6 12 6 10.4 6 8.5S7.6 5 9.5 5H11m8-2H9.5C6.5 3 4 5.5 4 8.5S6.5 14 9.5 14H11v7h2V5h2v16h2V5h2V3z" /></SVG>,
 
 	category: 'common',
 
@@ -195,17 +199,14 @@ export const settings = {
 				return <RawHTML>{ attributes.content }</RawHTML>;
 			},
 			migrate( attributes ) {
-				return {
-					...attributes,
-					content: create( { html: attributes.content } ),
-				};
+				return attributes;
 			},
 		},
 	],
 
 	merge( attributes, attributesToMerge ) {
 		return {
-			content: concat( attributes.content, attributesToMerge.content ),
+			content: attributes.content + attributesToMerge.content,
 		};
 	},