+++ packages/editor/src/components/post-publish-panel/postpublish.js
@@ -6,8 +6,8 @@ import { get } from 'lodash';
 /**
  * WordPress Dependencies
  */
-import { PanelBody, Button, ClipboardButton } from '@wordpress/components';
-import { __ } from '@wordpress/i18n';
+import { PanelBody, Button, ClipboardButton, TextControl } from '@wordpress/components';
+import { __, sprintf } from '@wordpress/i18n';
 import { Component, Fragment } from '@wordpress/element';
 import { withSelect } from '@wordpress/data';
 
@@ -49,14 +49,12 @@ class PostPublishPanelPostpublish extends Component {
 
 	render() {
 		const { children, isScheduled, post, postType } = this.props;
+		const postLabel = get( postType, [ 'labels', 'singular_name' ] );
 		const viewPostLabel = get( postType, [ 'labels', 'view_item' ] );
 
 		const postPublishNonLinkHeader = isScheduled ?
 			<Fragment>{ __( 'is now scheduled. It will go live on' ) } <PostScheduleLabel />.</Fragment> :
 			__( 'is now live.' );
-		const postPublishBodyText = isScheduled ?
-			__( 'The post address will be:' ) :
-			__( 'What’s next?' );
 
 		return (
 			<div className="post-publish-panel__postpublish">
@@ -64,13 +62,18 @@ class PostPublishPanelPostpublish extends Component {
 					<a href={ post.link }>{ post.title || __( '(no title)' ) }</a> { postPublishNonLinkHeader }
 				</PanelBody>
 				<PanelBody>
-					<div><strong>{ postPublishBodyText }</strong></div>
-					<input
-						className="post-publish-panel__postpublish-link-input"
+					<p className="post-publish-panel__postpublish-subheader">
+						<strong>{ __( 'What’s next?' ) }</strong>
+					</p>
+					<TextControl
+						className="post-publish-panel__postpublish-post-address"
 						readOnly
+						label={ sprintf(
+							/* translators: %s: post type singular name */
+							__( '%s address' ), postLabel
+						) }
 						value={ post.link }
 						onFocus={ this.onSelectInput }
-						type="text"
 					/>
 					<div className="post-publish-panel__postpublish-buttons">
 						{ ! isScheduled && (