+++ packages/editor/src/components/rich-text/index.js
@@ -9,6 +9,7 @@ import {
 	isNil,
 	noop,
 	isEqual,
+	omit,
 } from 'lodash';
 import memize from 'memize';
 
@@ -784,12 +785,14 @@ export class RichText extends Component {
 			const record = this.formatToValue( value );
 
 			if ( this.isActive() ) {
-				const length = getTextContent( prevProps.value ).length;
+				const prevRecord = this.formatToValue( prevProps.value );
+				const length = getTextContent( prevRecord ).length;
 				record.start = length;
 				record.end = length;
 			}
 
 			this.applyRecord( record );
+			this.savedContent = value;
 		}
 	}
 
@@ -939,7 +942,8 @@ export class RichText extends Component {
 
 RichText.defaultProps = {
 	formattingControls: FORMATTING_CONTROLS.map( ( { format } ) => format ),
-	format: 'rich-text',
+	format: 'string',
+	value: '',
 };
 
 const RichTextContainer = compose( [
@@ -987,40 +991,39 @@ const RichTextContainer = compose( [
 	withSafeTimeout,
 ] )( RichText );
 
-RichTextContainer.Content = ( { value, format, tagName: Tag, multiline, ...props } ) => {
+RichTextContainer.Content = ( { value, tagName: Tag, multiline: MultilineTag, ...props } ) => {
 	let html = value;
 
 	// Handle deprecated `children` and `node` sources.
 	if ( Array.isArray( value ) ) {
 		html = children.toHTML( value );
-	} else if ( format !== 'string' ) {
-		html = toHTMLString( value, multiline );
+	}
+
+	if ( ! html && MultilineTag ) {
+		html = `<${ MultilineTag }></${ MultilineTag }>`;
 	}
 
 	const content = <RawHTML>{ html }</RawHTML>;
 
 	if ( Tag ) {
-		return <Tag { ...props }>{ content }</Tag>;
+		return <Tag { ...omit( props, [ 'format' ] ) }>{ content }</Tag>;
 	}
 
 	return content;
 };
 
-RichTextContainer.isEmpty = ( value ) => {
+RichTextContainer.isEmpty = ( value = '' ) => {
 	// Handle deprecated `children` and `node` sources.
 	if ( Array.isArray( value ) ) {
 		return ! value || value.length === 0;
 	}
 
-	if ( typeof value === 'string' ) {
-		return value.length === 0;
-	}
-
-	return isEmpty( value );
+	return value.length === 0;
 };
 
 RichTextContainer.Content.defaultProps = {
-	format: 'rich-text',
+	format: 'string',
+	value: '',
 };
 
 export default RichTextContainer;