+++ packages/rich-text/src/apply-format.js
@@ -1,4 +1,10 @@
 /**
+ * External dependencies
+ */
+
+import { find } from 'lodash';
+
+/**
  * Internal dependencies
  */
 
@@ -24,15 +30,37 @@ export function applyFormat(
 ) {
 	const newFormats = formats.slice( 0 );
 
-	for ( let index = startIndex; index < endIndex; index++ ) {
-		if ( formats[ index ] ) {
-			const newFormatsAtIndex = formats[ index ].filter( ( { type } ) => type !== format.type );
-			newFormatsAtIndex.push( format );
-			newFormats[ index ] = newFormatsAtIndex;
-		} else {
-			newFormats[ index ] = [ format ];
+	// If the selection is collapsed, expand start and end to the edges of the
+	// format.
+	if ( startIndex === endIndex ) {
+		const startFormat = find( newFormats[ startIndex ], { type: format.type } );
+
+		while ( find( newFormats[ startIndex ], startFormat ) ) {
+			applyFormats( newFormats, startIndex, format );
+			startIndex--;
+		}
+
+		endIndex++;
+
+		while ( find( newFormats[ endIndex ], startFormat ) ) {
+			applyFormats( newFormats, endIndex, format );
+			endIndex++;
+		}
+	} else {
+		for ( let index = startIndex; index < endIndex; index++ ) {
+			applyFormats( newFormats, index, format );
 		}
 	}
 
 	return normaliseFormats( { formats: newFormats, text, start, end } );
+}
+
+function applyFormats( formats, index, format ) {
+	if ( formats[ index ] ) {
+		const newFormatsAtIndex = formats[ index ].filter( ( { type } ) => type !== format.type );
+		newFormatsAtIndex.push( format );
+		formats[ index ] = newFormatsAtIndex;
+	} else {
+		formats[ index ] = [ format ];
+	}
 }