+++ packages/block-library/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.1.0 (2018-10-10)
+
+### New Features
+
+- Include the classic block if `wp.oldEditor` is defined.
+- Include the HTML block.
+
 ## 2.0.0 (2018-09-05)
 
 ### Breaking Change